Move git-specific repo_url logic into git plugin
