+++ lib/capistrano/dsl/paths.rb
@@ -36,20 +36,7 @@ module Capistrano
       end
 
       def repo_url
-        require "cgi"
-        require "uri"
-        if fetch(:git_http_username) && fetch(:git_http_password)
-          URI.parse(fetch(:repo_url)).tap do |repo_uri|
-            repo_uri.user     = fetch(:git_http_username)
-            repo_uri.password = CGI.escape(fetch(:git_http_password))
-          end.to_s
-        elsif fetch(:git_http_username)
-          URI.parse(fetch(:repo_url)).tap do |repo_uri|
-            repo_uri.user = fetch(:git_http_username)
-          end.to_s
-        else
-          fetch(:repo_url)
-        end
+        fetch(:repo_url)
       end
 
       def repo_path
